# backend/src/main/resources/openapi/cms-api.yaml
openapi: 3.0.3
info:
  title: CMS Internal Knowledge Portal API
  version: 0.1.0
  description: 사내 지식·문서·공지 CMS와 사용자 포털을 위한 REST API 계약
servers:
  - url: https://cms.example.com
    description: production
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/content/{contentId}:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    get:
      tags: [Content]
      operationId: getContentForAdmin
      summary: CMS 콘텐츠 상세 조회
      responses:
        "200":
          description: 콘텐츠 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Content]
      operationId: updateContent
      summary: 콘텐츠 초안 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdateRequest"
      responses:
        "200":
          description: 수정된 콘텐츠
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 최신 버전 충돌
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Content]
      operationId: deleteDraftContent
      summary: 발행 전 초안 삭제
      responses:
        "204":
          description: 삭제됨
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/portal/home:
    get:
      tags: [Portal]
      operationId: getPortalHome
      summary: 포털 홈 데이터 조회
      responses:
        "200":
          description: 포털 홈 섹션 데이터
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalHomeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/portal/search:
    get:
      tags: [Portal]
      operationId: searchPortalContent
      summary: 포털 통합 검색
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            maxLength: 200
        - name: contentType
          in: query
          schema:
            $ref: "#/components/schemas/ContentType"
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: tag
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 사용자 권한이 적용된 검색 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
  /api/v1/portal/content/{contentId}:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    get:
      tags: [Portal]
      operationId: getPortalContent
      summary: 포털 콘텐츠 상세 조회
      responses:
        "200":
          description: 공개 대상 검증 후 콘텐츠 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portal/bookmarks:
    get:
      tags: [Portal]
      operationId: listBookmarks
      summary: 내 즐겨찾기 목록 조회
      responses:
        "200":
          description: 즐겨찾기 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentSummary"
    post:
      tags: [Portal]
      operationId: addBookmark
      summary: 즐겨찾기 추가
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkRequest"
      responses:
        "201":
          description: 즐겨찾기 추가됨
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/portal/bookmarks/{contentId}:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    delete:
      tags: [Portal]
      operationId: removeBookmark
      summary: 즐겨찾기 해제
      responses:
        "204":
          description: 즐겨찾기 해제됨
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 요청 값 검증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ContentType:
      type: string
      enum: [ARTICLE, DOCUMENT, NOTICE]
    ContentStatus:
      type: string
      enum: [DRAFT, IN_REVIEW, REJECTED, APPROVED, SCHEDULED, PUBLISHED, ARCHIVED, EXPIRED]
    RoleCode:
      type: string
      enum: [ADMIN, EDITOR, REVIEWER, EMPLOYEE, VIEWER]
    UserProfile:
      type: object
      required: [id, email, displayName, roles, permissions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        department:
          $ref: "#/components/schemas/Department"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleCode"
        permissions:
          type: array
          items:
            type: string
    Department:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    Tag:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
...
    AudienceRule:
      type: object
      required: [audienceType]
      properties:
        audienceType:
          type: string
          enum: [ALL_EMPLOYEES, DEPARTMENT, ROLE, USER]
        audienceRefId:
          type: string
          nullable: true
          description: "`ALL_EMPLOYEES` 외 대상 참조 식별자"
    Attachment:
      type: object
      required: [id, fileName, mimeType, sizeBytes, scanStatus]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        scanStatus:
          type: string
          enum: [PENDING, CLEAN, INFECTED, FAILED]
        downloadUrl:
          type: string
          format: uri
          nullable: true
    ContentSummary:
      type: object
      required: [id, contentType, status, title, summary, category, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        contentType:
          $ref: "#/components/schemas/ContentType"
        status:
          $ref: "#/components/schemas/ContentStatus"
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        isImportant:
          type: boolean
        requiresAcknowledgement:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
    ContentDetail:
      allOf:
        - $ref: "#/components/schemas/ContentSummary"
        - type: object
          required: [body, author, audiences, versionNumber]
          properties:
            body:
              type: string
              description: 리치 텍스트 본문 payload
            author:
              $ref: "#/components/schemas/UserProfile"
            audiences:
              type: array
              items:
                $ref: "#/components/schemas/AudienceRule"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/Attachment"
            versionNumber:
              type: integer
            scheduledAt:
              type: string
              format: date-time
              nullable: true
            expiresAt:
              type: string
              format: date-time
              nullable: true
    ContentCreateRequest:
      type: object
      required: [contentType, title, body, categoryId, audiences]
      properties:
        contentType:
          $ref: "#/components/schemas/ContentType"
        title:
          type: string
          maxLength: 150
        summary:
          type: string
          maxLength: 300
        body:
          type: string
        categoryId:
          type: string
          format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid
        audiences:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AudienceRule"
        isImportant:
          type: boolean
          default: false
        requiresAcknowledgement:
          type: boolean
          default: false
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ContentUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/ContentCreateRequest"
        - type: object
          required: [baseVersionNumber, changeSummary]
          properties:
            baseVersionNumber:
              type: integer
            changeSummary:
              type: string
              maxLength: 500
...
    PortalHomeResponse:
      type: object
      required: [requiredNotices, latestUpdates, popularContent, categoryShortcuts]
      properties:
        requiredNotices:
          type: array
          items:
            $ref: "#/components/schemas/NoticeSummary"
        latestUpdates:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
        popularContent:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
        categoryShortcuts:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    SearchResponse:
      type: object
      required: [items, page, resultCount]
      properties:
        query:
          type: string
        resultCount:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
        page:
          $ref: "#/components/schemas/PageInfo"
    BookmarkRequest:
      type: object
      required: [contentId]
      properties:
        contentId:
          type: string
          format: uuid
    NoticeSummary:
      allOf:
        - $ref: "#/components/schemas/ContentSummary"
        - type: object
          properties:
            acknowledgementStatus:
              type: string
              enum: [PENDING, ACKNOWLEDGED, EXEMPTED]
            acknowledgedAt:
              type: string
              format: date-time
              nullable: true
...
    PageInfo:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string